Refactor gql queries and tests
