+++ src/entrypoint.module.ts
@@ -21,4 +21,6 @@ export function run(
 export { BackportResponse } from './main';
 export { ConfigOptions } from './options/ConfigOptions';
 export { getTargetBranchForLabel } from './services/github/v4/getTargetBranchesFromLabels';
-export { fetchMergedPullRequests } from './services/github/v4/fetchMergedPullRequests';
+export { fetchCommitsByAuthor } from './services/github/v4/fetchCommits/fetchCommitsByAuthor';
+export { fetchCommitByPullNumber } from './services/github/v4/fetchCommits/fetchCommitByPullNumber';
+export { fetchCommitBySha } from './services/github/v4/fetchCommits/fetchCommitBySha';
