+++ src/services/github/v4/fetchCommits/fetchPullRequestBySearchQuery.private.test.ts
@@ -1,5 +1,5 @@
-import { ValidConfigOptions } from '../../../options/options';
-import { getDevAccessToken } from '../../../test/private/getDevAccessToken';
+import { ValidConfigOptions } from '../../../../options/options';
+import { getDevAccessToken } from '../../../../test/private/getDevAccessToken';
 import { fetchPullRequestBySearchQuery } from './fetchPullRequestBySearchQuery';
 
 describe('fetchPullRequestBySearchQuery', () => {
