--- src/services/github/v4/fetchDefaultBranch.ts
@@ -1,32 +1,0 @@
-import { ValidConfigOptions } from '../../../options/options';
-import { apiRequestV4 } from './apiRequestV4';
-
-export async function fetchDefaultBranch(options: ValidConfigOptions) {
-  const { accessToken, githubApiBaseUrlV4, repoName, repoOwner } = options;
-  const query = /* GraphQL */ `
-    query DefaultBranch($repoOwner: String!, $repoName: String!) {
-      repo: repository(owner: $repoOwner, name: $repoName) {
-        defaultBranchRef {
-          name
-        }
-      }
-    }
-  `;
-
-  const res = await apiRequestV4<DefaultBranchResoonse>({
-    githubApiBaseUrlV4,
-    accessToken,
-    query,
-    variables: { repoOwner, repoName },
-  });
-
-  return res.repo.defaultBranchRef.name;
-}
-
-export interface DefaultBranchResoonse {
-  repo: {
-    defaultBranchRef: {
-      name: string;
-    };
-  };
-}
