+++ src/services/github/v4/mocks/commitsByAuthorMock.ts
@@ -1,6 +1,6 @@
-import { CommitByAuthorResponse } from '../fetchCommitsByAuthor';
+import { CommitByAuthorResponse } from '../fetchCommits/fetchCommitsByAuthor';
 
-export const commitsWithPullRequestsMock: CommitByAuthorResponse = {
+export const commitsByAuthorMock: CommitByAuthorResponse = {
   repository: {
     ref: {
       target: {
@@ -8,6 +8,10 @@ export const commitsWithPullRequestsMock: CommitByAuthorResponse = {
           edges: [
             {
               node: {
+                repository: {
+                  name: 'kibana',
+                  owner: { login: 'elastic' },
+                },
                 committedDate: '2021-12-24T00:00:00Z',
                 oid: '2e63475c483f7844b0f2833bc57fdee32095bacb',
                 message: 'Add 👻',
@@ -18,6 +22,10 @@ export const commitsWithPullRequestsMock: CommitByAuthorResponse = {
             },
             {
               node: {
+                repository: {
+                  name: 'kibana',
+                  owner: { login: 'elastic' },
+                },
                 committedDate: '2021-12-23T00:00:00Z',
                 oid: 'f3b618b9421fdecdb36862f907afbdd6344b361d',
                 message: 'Add witch (#85)',
@@ -28,17 +36,6 @@ export const commitsWithPullRequestsMock: CommitByAuthorResponse = {
                         baseRefName: 'master',
                         labels: {
                           nodes: [{ name: 'my-label-b' }],
-                        },
-                        mergeCommit: {
-                          committedDate: '2021-01-01T00:00:00',
-                          oid: 'f3b618b9421fdecdb36862f907afbdd6344b361d',
-                          message: 'Add witch (#85)',
-                        },
-                        repository: {
-                          name: 'kibana',
-                          owner: {
-                            login: 'elastic',
-                          },
                         },
                         number: 85,
                         timelineItems: {
@@ -52,6 +49,10 @@ export const commitsWithPullRequestsMock: CommitByAuthorResponse = {
             },
             {
               node: {
+                repository: {
+                  name: 'kibana',
+                  owner: { login: 'elastic' },
+                },
                 committedDate: '2021-12-22T00:00:00Z',
                 oid: '79cf18453ec32a4677009dcbab1c9c8c73fc14fe',
                 message:
@@ -63,25 +64,14 @@ export const commitsWithPullRequestsMock: CommitByAuthorResponse = {
                         labels: {
                           nodes: [{ name: 'my-label-a' }],
                         },
-                        mergeCommit: {
-                          committedDate: '2021-01-01T00:00:00',
-                          oid: '79cf18453ec32a4677009dcbab1c9c8c73fc14fe',
-                          message:
-                            'Add SF mention (#80)\n\n* Add SF mention\r\n\r\n* Add several emojis!',
-                        },
-                        repository: {
-                          name: 'kibana',
-                          owner: {
-                            login: 'elastic',
-                          },
-                        },
+
                         baseRefName: 'master',
                         number: 80,
                         timelineItems: {
                           edges: [
                             {
                               node: {
-                                source: {
+                                targetPullRequest: {
                                   __typename: 'PullRequest',
                                   title: 'some title',
                                   state: 'MERGED',
@@ -91,7 +81,14 @@ export const commitsWithPullRequestsMock: CommitByAuthorResponse = {
                                     edges: [
                                       {
                                         node: {
-                                          commit: {
+                                          targetCommit: {
+                                            repository: {
+                                              name: 'kibana',
+                                              owner: { login: 'elastic' },
+                                            },
+                                            committedDate:
+                                              '2021-12-23T00:00:00Z',
+                                            oid: 'abc',
                                             message:
                                               'Add SF mention (#80)\n\n* Add SF mention\r\n\r\n* Add several emojis!',
                                           },
@@ -112,6 +109,10 @@ export const commitsWithPullRequestsMock: CommitByAuthorResponse = {
             },
             {
               node: {
+                repository: {
+                  name: 'kibana',
+                  owner: { login: 'elastic' },
+                },
                 committedDate: '2021-12-21T00:00:00Z',
                 oid: '3827bbbaf39914eda4f02f6940189844375fd097',
                 message: 'Add backport config',
@@ -122,6 +123,10 @@ export const commitsWithPullRequestsMock: CommitByAuthorResponse = {
             },
             {
               node: {
+                repository: {
+                  name: 'kibana',
+                  owner: { login: 'elastic' },
+                },
                 committedDate: '2021-12-20T00:00:00Z',
                 oid: '5ea0da550ac191029459289d67f99ad7d310812b',
                 message: 'Initial commit',
