+++ src/ui/cherrypickAndCreatePullRequest.test.ts
@@ -5,9 +5,9 @@ import * as childProcess from '../services/child-process-promisified';
 import * as logger from '../services/logger';
 import * as prompts from '../services/prompts';
 import { ExecError } from '../test/ExecError';
-import { Commit } from '../types/Commit';
 import { PromiseReturnType } from '../types/PromiseReturnType';
 import { SpyHelper } from '../types/SpyHelper';
+import { Commit } from '../types/commitWithAssociatedPullRequests';
 import { cherrypickAndCreateTargetPullRequest } from './cherrypickAndCreateTargetPullRequest';
 
 describe('cherrypickAndCreateTargetPullRequest', () => {
